+++ CHANGELOG.rst
@@ -26,6 +26,8 @@
 - Throw a more intelligle error during simulation if a topology is produced
   where the time of a parent is equal to the time of the child. (#570, #87).
 
+- Pickle supported in the TableCollection object. (#574, #577).
+
 **Deprecated**:
 
 - The ``filter_zero_mutation_sites`` parameter for simplify has been deprecated